@@ -16,7 +16,7 @@ Gem::Specification.new do |s|
 
   s.files = Dir["{app,config,db,lib}/**/*", "MIT-LICENSE", "Rakefile", "README.md"]
 
-  s.add_dependency "rails", ">= 5.0.0.1"
+  s.add_dependency "rails", ">= 6.0.0.1"
   s.add_dependency "lodash-rails", ">= 4.15.0"
 
   s.add_development_dependency "sqlite3"
