@@ -1,3 +1,3 @@
 module ActionCableNotifications
-  VERSION = '0.1.37'
+  VERSION = '0.2.0'
 end
